修复稍后再看存在的用户id无法解析的隐患

+++ src/Models/Models.BiliBili/Core/PublisherInfo.cs
@@ -13,7 +13,7 @@ namespace Bili.Models.BiliBili
         /// 视频发布者的Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "mid", Required = Required.Default)]
-        public int Mid { get; set; }
+        public long Mid { get; set; }
 
         /// <summary>
         /// 视频发布者.
